# Switches embedded server from Tomcat to Undertow