--- a/zipkin-autoconfigure/metrics-prometheus/src/main/resources/META-INF/spring.factories
+++ b/zipkin-autoconfigure/metrics-prometheus/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-zipkin.autoconfigure.metrics.PrometheusMetricsAutoConfiguration
+zipkin.autoconfigure.prometheus.ZipkinPrometheusMetricsAutoConfiguration