--- a/zipkin-server/src/main/java/zipkin/server/ZipkinServerConfiguration.java
+++ b/zipkin-server/src/main/java/zipkin/server/ZipkinServerConfiguration.java
@@ -14,7 +14,6 @@
 package zipkin.server;
 
 import com.github.kristofa.brave.Brave;
-import java.util.Arrays;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -23,15 +22,13 @@ import org.springframework.boot.actuate.health.HealthAggregator;
 import org.springframework.boot.actuate.metrics.buffer.CounterBuffers;
 import org.springframework.boot.actuate.metrics.buffer.GaugeBuffers;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.context.embedded.undertow.UndertowEmbeddedServletContainerFactory;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Condition;
 import org.springframework.context.annotation.ConditionContext;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.type.AnnotatedTypeMetadata;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
 import zipkin.collector.CollectorMetrics;
 import zipkin.collector.CollectorSampler;
 import zipkin.internal.V2StorageComponent;
@@ -39,14 +36,23 @@ import zipkin.server.brave.TracedStorageComponent;
 import zipkin.storage.StorageComponent;
 import zipkin2.storage.InMemoryStorage;
 
-import static java.util.Arrays.asList;
-
 @Configuration
 public class ZipkinServerConfiguration {
 
   /** Registers health for any components, even those not in this jar. */
   @Bean ZipkinHealthIndicator zipkinHealthIndicator(HealthAggregator healthAggregator) {
     return new ZipkinHealthIndicator(healthAggregator);
+  }
+
+  @Bean public UndertowEmbeddedServletContainerFactory embeddedServletContainerFactory(
+    @Value("${zipkin.query.allowed-origins:*}") String allowedOrigins
+  ) {
+    UndertowEmbeddedServletContainerFactory factory = new UndertowEmbeddedServletContainerFactory();
+    CorsHandler cors = new CorsHandler(allowedOrigins);
+    factory.addDeploymentInfoCustomizers(
+      info -> info.addInitialHandlerChainWrapper(cors)
+    );
+    return factory;
   }
 
   @Bean
@@ -57,14 +63,15 @@ public class ZipkinServerConfiguration {
 
   @Bean
   @ConditionalOnMissingBean(CollectorMetrics.class)
-  CollectorMetrics metrics(Optional<CounterBuffers> counterBuffers, Optional<GaugeBuffers> gaugeBuffers) {
+  CollectorMetrics metrics(Optional<CounterBuffers> counterBuffers,
+    Optional<GaugeBuffers> gaugeBuffers) {
     // it is not guaranteed that BufferCounterService/CounterBuffers will be used,
     // for ex., com.datastax.cassandra:cassandra-driver-core brings com.codahale.metrics.MetricRegistry
     // and as result DropwizardMetricServices is getting instantiated instead of standard Java8 BufferCounterService.
     // On top of it Cassandra driver heavily relies on Dropwizard metrics and manually excluding it from pom.xml is not an option.
     // MetricsDropwizardAutoConfiguration can be manually excluded either, as Cassandra metrics won't be recorded.
     return new ActuateCollectorMetrics(counterBuffers.orElse(new CounterBuffers()),
-                                       gaugeBuffers.orElse(new GaugeBuffers()));
+      gaugeBuffers.orElse(new GaugeBuffers()));
   }
 
   @Configuration
@@ -87,19 +94,6 @@ public class ZipkinServerConfiguration {
       }
       return bean;
     }
-  }
-
-  @Bean
-  @ConditionalOnMissingBean(CorsFilter.class)
-  CorsFilter corsFilter(@Value("${zipkin.query.allowed-origins:*}") String allowedOrigins) {
-    CorsConfiguration configuration = new CorsConfiguration();
-    configuration.setAllowedOrigins(asList(allowedOrigins.split(",")));
-    configuration.setAllowedMethods(asList("GET", "POST"));
-    configuration.setAllowCredentials(false);
-    UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-    source.registerCorsConfiguration("/api/**", configuration);
-    source.registerCorsConfiguration("/zipkin/api/**", configuration);
-    return new CorsFilter(source);
   }
 
   /**
@@ -128,7 +122,7 @@ public class ZipkinServerConfiguration {
     @Override public boolean matches(ConditionContext condition, AnnotatedTypeMetadata ignored) {
       String storageType = condition.getEnvironment().getProperty("zipkin.storage.type");
       if (storageType == null) return true;
-      storageType  = storageType.trim();
+      storageType = storageType.trim();
       if (storageType.isEmpty()) return true;
       return storageType.equals("mem");
     }