This switches from Tomcat to Undertow, then MVC to raw Undertow handler to process http messages (for apps sending data directly to zipkin).

The test uses a real backend even if local elasticsearch. The test results in a lot more data than my local ES can swallow, so it is also a test of how fast we can drop data, and how well our rate-limiting works.

So, after this change, the same zipkin server against local elasticsearch can swallow 3x more traffic than before.

The below results were a best of 5 for before and after.. tests were interleaved to ensure elasticsearch wasn't skewing.

```
# before
Running 30s test @ http://localhost:9411/api/v2/spans
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    14.50ms   37.19ms 527.63ms   93.43%
    Req/Sec     2.02k     0.94k    6.66k    71.75%
  Latency Distribution
     50%    4.20ms
     75%    8.61ms
     90%   24.53ms
     99%  198.98ms
  212147 requests in 30.05s, 24.92MB read
  Socket errors: connect 0, read 1, write 0, timeout 19
Requests/sec:   7058.96
Transfer/sec:    849.19KB

# after
Running 30s test @ http://localhost:9411/api/v2/spans
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.28ms   33.11ms 388.07ms   90.37%
    Req/Sec     6.43k     2.86k   14.55k    64.65%
  Latency Distribution
     50%    3.03ms
     75%   14.59ms
     90%   47.73ms
     99%  155.01ms
  761153 requests in 30.05s, 76.22MB read
Requests/sec:  25332.61
Transfer/sec:      2.54MB
```

Here's the command which starts a server and uses elasticsearch local storage:
```bash
$ STORAGE_TYPE=elasticsearch java -jar ./zipkin-server/target/zipkin-server-*-exec.jar
$ wrk -t4 -c128 -d30s http://localhost:9411/api/v2/spans -s zipkin-json.lua --latency
```

Here's the script:
```lua
wrk.method = "POST"
wrk.headers["Content-Type"] = "application/json"
wrk.body   = '[{"traceId":"4d1e00c0db9010db86154a4ba6e91385","parentId":"86154a4ba6e91385","id":"4d1e00c0db9010db","kind":"CLIENT","name":"get","timestamp":1472470996199000,"duration":207000,"localEndpoint":{"serviceName":"frontend","ipv6":"7::0.128.128.127"},"remoteEndpoint":{"serviceName":"backend","ipv4":"192.168.99.101","port":9000},"annotations":[{"timestamp":1472470996238000,"value":"foo"},{"timestamp":1472470996403000,"value":"bar"}],"tags":{"http.path":"/api","clnt/finagle.version":"6.45.0"}}]'
```

I looked at the jar, and it is roughly the same size as before. Since
undertow uses the relatively obscure xnio library, it isn't likely to
conflict with more often used things like netty.