@adriancole 2017-11-20
best run of several (locally) with tomcat:
```
Running 30s test @ http://localhost:9411/api/v2/spans
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    11.23ms   18.08ms 400.13ms   93.27%
    Req/Sec     4.23k   496.83     7.75k    77.38%
  Latency Distribution
     50%    7.10ms
     75%    8.86ms
     90%   16.97ms
     99%  100.63ms
  504453 requests in 30.05s, 59.26MB read
  Socket errors: connect 0, read 11, write 0, timeout 0
Requests/sec:  16789.47
Transfer/sec:      1.97MB
```

Best run of several (locally) with undertow
```
Running 30s test @ http://localhost:9411/api/v2/spans
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     9.79ms   16.26ms 421.10ms   93.06%
    Req/Sec     5.05k   547.20     8.77k    73.58%
  Latency Distribution
     50%    5.51ms
     75%    7.37ms
     90%   17.48ms
     99%   87.97ms
  604521 requests in 30.10s, 89.36MB read
Requests/sec:  20085.30
Transfer/sec:      2.97MB
```