@adriancole 2017-11-20
cc @anuraaga @shakuzen @denyska FYI a small bump in http ingest speed for basically no effort