@adriancole 2017-11-21
I'm going to revise this to yank the MVC controller as there's a large amount of latency added there. Here's raw undertow (still wired with spring boot), which has dramatically better latency characteristics. The impact is we can chop off several milliseconds off the better requests and an order of magnitude off the worse ones. This is particularly helpful when collector-tier sampling is in place as it means we can skip data much quicker under a surge.

```bash
$ wrk -t4 -c128 -d30s http://localhost:9411/api/v2/spans -s zipkin-json.lua --latency
Running 30s test @ http://localhost:9411/api/v2/spans
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    11.23ms   18.08ms 400.13ms   93.27%
    Req/Sec     4.23k   496.83     7.75k    77.38%
  Latency Distribution
     50%    7.10ms
     75%    8.86ms
     90%   16.97ms
     99%  100.63ms
  504453 requests in 30.05s, 59.26MB read
  Socket errors: connect 0, read 11, write 0, timeout 0
Requests/sec:  16789.47
Transfer/sec:      1.97MB
```