@adriancole 2017-11-22
ok code is up. I changed the test to use a real backend even if local elasticsearch. The test results in a lot more data than ES can swallow, so it is also a test of how fast we can drop data, and how well our rate-limiting works.

So, after this change, the same zipkin server against local elasticsearch can swallow 3x more traffic than before.

The below results were a best of 5 for before and after.. tests were interleaved to ensure elasticsearch wasn't skewing.

```
# before
Running 30s test @ http://localhost:9411/api/v2/spans
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    14.50ms   37.19ms 527.63ms   93.43%
    Req/Sec     2.02k     0.94k    6.66k    71.75%
  Latency Distribution
     50%    4.20ms
     75%    8.61ms
     90%   24.53ms
     99%  198.98ms
  212147 requests in 30.05s, 24.92MB read
  Socket errors: connect 0, read 1, write 0, timeout 19
Requests/sec:   7058.96
Transfer/sec:    849.19KB

# after
Running 30s test @ http://localhost:9411/api/v2/spans
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    16.28ms   33.11ms 388.07ms   90.37%
    Req/Sec     6.43k     2.86k   14.55k    64.65%
  Latency Distribution
     50%    3.03ms
     75%   14.59ms
     90%   47.73ms
     99%  155.01ms
  761153 requests in 30.05s, 76.22MB read
Requests/sec:  25332.61
Transfer/sec:      2.54MB
```