@adriancole 2017-11-22
PS there's clearly more possible here, just looking for a cheap win