@adriancole 2017-11-23
ok I rejigged the cors filter into its own commit (tested it with the zipkin-js browser example). I'm going to test a proxied zipkin ui (nginx) before merging