@adriancole 2017-11-23
noticed our metrics were double-counting so fixed that before proceeding. I want to watch our prometheus dashboard before and after..