@adriancole 2017-11-23
yanked the servlet filter for request duration as we don't need it anymore