@adriancole 2017-11-23
TL;DR; this is without a doubt far better especially in my container environment.

I tested in docker. This version does quite a lot better in terms of resources used, time in GC, etc.. Using elasticsearch5 image and grafana etc

Before: best is 587 requests/second with P99 1second
```
Running 30s test @ http://192.168.99.100:9411/api/v2/spans
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   251.37ms  223.66ms   1.96s    70.74%
    Req/Sec   148.22     90.62   540.00     66.30%
  Latency Distribution
     50%  210.87ms
     75%  357.93ms
     90%  552.59ms
     99%  995.82ms
  17665 requests in 30.06s, 2.85MB read
Requests/sec:    587.62
Transfer/sec:     97.06KB
```

After: best is 2.5K requests/second with P99 190ms
```
Running 30s test @ http://192.168.99.100:9411/api/v2/spans
  4 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    53.76ms   35.06ms 346.09ms   79.44%
    Req/Sec   640.58    220.66     1.73k    75.34%
  Latency Distribution
     50%   48.23ms
     75%   64.04ms
     90%   90.27ms
     99%  191.56ms
  76292 requests in 30.01s, 10.99MB read
Requests/sec:   2541.89
Transfer/sec:    374.83KB
```

All things monitored in graphite were measurably better including load, thread count, time in GC

<img width="503" alt="screen shot 2017-11-23 at 10 41 43 pm" src="https://user-images.githubusercontent.com/64215/33178408-0ad8f152-d0a1-11e7-9c65-5813e6047904.png">
<img width="496" alt="screen shot 2017-11-23 at 10 47 40 pm" src="https://user-images.githubusercontent.com/64215/33178412-0e46440c-d0a1-11e7-860c-1f9e6d4a7175.png">
<img width="461" alt="screen shot 2017-11-23 at 10 41 34 pm" src="https://user-images.githubusercontent.com/64215/33178429-181ad97a-d0a1-11e7-8679-677f779f963e.png">
<img width="497" alt="screen shot 2017-11-23 at 10 47 25 pm" src="https://user-images.githubusercontent.com/64215/33178426-13ece3e8-d0a1-11e7-87bb-022b5fc67ada.png">
<img width="498" alt="screen shot 2017-11-23 at 10 41 28 pm" src="https://user-images.githubusercontent.com/64215/33178441-20f93a28-d0a1-11e7-820c-a8ebd897d58d.png">
<img width="502" alt="screen shot 2017-11-23 at 10 47 31 pm" src="https://user-images.githubusercontent.com/64215/33178436-1d3fea58-d0a1-11e7-9666-ca8bad04accc.png">